requirements.txt
Django==1.6.2
South==1.0.1
argparse==1.2.1
django-braces==1.4.0
django-crispy-forms==1.4.0
six==1.8.0
wsgiref==0.1.2
